backend/requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
requests
python-dotenv
